+++ src/app/carteira/page.tsx
@@ -3,8 +3,8 @@
 import * as Dialog from "@radix-ui/react-dialog";
 import { useState } from "react";
 
-import { Button } from "@/components/Button";
-import { Modal } from "@/components/Modal";
+import { Button } from "@/components";
+import { Modal } from "@/components";
 import { RegisterWallet } from "./components/RegisterWallet/register-wallet.component";
 
 export function Carteira() {
